@@ -14,16 +14,17 @@
     "nano-ipfs-store": "latest",
     "prop-types": "latest",
     "ramda": "^0.25.0",
-    "react": "^16.6.1",
+    "react": "16.7.0-alpha.0",
     "react-codemirror": "^1.0.0",
-    "react-dom": "^16.6.1",
+    "react-dom": "16.7.0-alpha.0",
     "react-redux": "^5.1.1",
     "react-router": "latest",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
     "redux": "^4.0.1",
     "redux-mock-store": "^1.5.3",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "react-test-renderer": "latest"
   },
   "devDependencies": {
     "enzyme": "latest",
